+++ spotinst/ocean_gke_launch_spec_strategy/consts.go
@@ -1,0 +1,8 @@
+package ocean_gke_launch_spec_strategy
+
+import "github.com/spotinst/terraform-provider-spotinst/spotinst/commons"
+
+const (
+	Strategy              commons.FieldName = "strategy"
+	PreemptiblePercentage commons.FieldName = "preemptible_percentage"
+)